python-crewai/requirements-api.txt
# API Requirements - MIT Tracking GraphQL + OpenAPI
# Adicionais à requirements.txt base

# FastAPI e dependências
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# GraphQL
strawberry-graphql[fastapi]>=0.215.0

# Middlewares e segurança
python-multipart>=0.0.6
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# Monitoramento e logging
structlog>=23.2.0
prometheus-client>=0.19.0

# Utilitários adicionais
aiofiles>=23.2.0
jinja2>=3.1.0  # Para templates se necessário